+++ addon/mixins/color-mixin.js
@@ -1,0 +1,5 @@
+import Ember from 'ember';
+
+export default Ember.Mixin.create({
+  classNameBindings: ['warn:md-warn', 'accent:md-accent', 'primary:md-primary']
+});