+++ tests/unit/components/paper-progress-linear-test.js
@@ -4,7 +4,7 @@ moduleForComponent('paper-progress-linear', 'Unit | Component | paper progress l
   // Specify the other units that are required for this test
   // needs: ['component:foo', 'helper:bar'],
   unit: true,
-  needs: ['service:browser-compatibility', 'service:sniffer']
+  needs: ['service:constants', 'service:sniffer']
 });
 
 test('it renders', function(assert) {
@@ -31,7 +31,7 @@ test('it sets transform based on value', function(assert) {
 
   var bar2 = this.$().find('.md-bar2')[0];
 
-  var bar2style = bar2.style[component.get('browserCompatibility.CSS.TRANSFORM')];
+  var bar2style = bar2.style[component.get('constants.CSS.TRANSFORM')];
 
   assert.equal(bar2style, 'translateX(-25%) scale(0.5, 1)', 'Transition set correctly');
 
@@ -40,7 +40,7 @@ test('it sets transform based on value', function(assert) {
 test('it sets transform based on buffer value', function(assert) {
   var component = this.subject({
     value: 50,
-    'md-buffer-value': 75
+    'buffer-value': 75
 
   });
 
@@ -48,7 +48,7 @@ test('it sets transform based on buffer value', function(assert) {
 
   var bar1 = this.$().find('.md-bar1')[0];
 
-  var bar1style = bar1.style[component.get('browserCompatibility.CSS.TRANSFORM')];
+  var bar1style = bar1.style[component.get('constants.CSS.TRANSFORM')];
 
   assert.equal(bar1style, 'translateX(-12.5%) scale(0.75, 1)', 'Buffer bar transition set correctly');
 });
@@ -56,14 +56,14 @@ test('it sets transform based on buffer value', function(assert) {
 test('it should not set transition in query mode', function(assert) {
   var component = this.subject({
     value: 80,
-    'md-mode': 'query'
+    'mode': 'query'
   });
 
   this.render();
 
   var bar2 = this.$().find('.md-bar2')[0];
 
-  var bar2style = bar2.style[component.get('browserCompatibility.CSS.TRANSFORM')];
+  var bar2style = bar2.style[component.get('constants.CSS.TRANSFORM')];
 
   assert.ok(!bar2style, 'Buffer bar not set');