+++ tests/unit/components/paper-switch-test.js
@@ -6,6 +6,7 @@ import {
 moduleForComponent('paper-switch', 'PaperToggleComponent', {
   // specify the other units that are required for this test
   // needs: ['component:foo', 'helper:bar']
+  unit: true
 });
 
 test('it renders', function(assert) {